Add sorter module
This is still in development, since the data current indexing is not
adapted to the sorting we choose (sorted by group before type, but
currently indexed by type), the post processor will only sort and log
its result without altering the actual data (so it won't break the API).

This is only useful to see the sort code is actually working, and can
be used in the latter implementation where the data is suitable.