Allow to pass a custom logger
    sassdoc(src, { logger: myLogger });

Before this was not working because the environment will throw when
seeing a `logger` key in the configuration because it's a reserved
attribute.

The `ensureLogger` function will now delete the `logger` config key to
avoid this problem.

Also the `checkLogger` function will provide friendly error handling,
and some tolerance over the given logger, allowing to not implement
non-critical `debug` method.