Graceful relative paths in CLI and config file
Before this, `dest` was always relative to CWD even when set from config
file, and if a config file was defined, `theme` was always relative to
it even when set via CLI.

Now for all CLI-passed environment properties, a "friend" `*Cli` prperty
is set to `true` (`themeCli`, `destCli`), so the environment loader can
determine whether it's relative to CWD or config file.

Also paths are now stored absolute, but are always displayed relative to
the CWD for coherent CLI output. If you set `theme: ../../theme` in
`foo/config.yml`, the displayed theme in CLI output will be `../theme`.

This fixes #362.