Fix group sorting
Since 2.0 data upgrade, the `group` key is no more an array of arrays,
it's just a flat array. The sorter was not up-to-date, but we did not
notice it since it was in fact sorting on the first letter of the group
name, which will look fine in most cases.

A test fixture was also wrong about this...